all: make Report return a Context, prepare lib to support APM (#3)
This commit alters Report to return a context, as well as changing the
way traceID's are generated. traceID's are also now referred to as
reportID's to avoid overloading the term.

In the future, it's likely that reportID's will be removed altogether,
since they *shouldn't* be necessary if receivers of Reports can receive
and modify a Context.