all: Add integration testing harness and some tests
This commit adds an integration test harness to the repository.

Individual tests consist of a directory under _integration-tests/tests
which contains a go program and a file, validation.json.

The go program will be instrumented by orchestrion, then run. The traces
output by that program go to a fake agent, which records the traces.

The validation.json file describes the expected structure of the traces
which should be produced by each test. This includes things like name,
resource, service, metadata and metrics.

Once the traces are recorded and the test program shuts down, the traces
recorded by the fake agent are checked against the expected ones
described in validation.json to ensure the test produced exactly what we
expected, in terms of the number of traces, the structure of those
traces, and the fields in the spans of the traces.