feat: decorate output of `toolexec [...] -V=full`
The go build toolchain calls tooling with `-V=full` to compute the full
build ID, allowing for invalidation of cached build artifacts when
tooling changes.

The contract for this option is to print a single line of content, which
will be embedded in the build ID. We are hence intercepting these in the
`-toolexec` flow and appending the Orchestrion version and ruleset
information to the output.