fix: `append-args` breaks on ellipsis
When an ellipsis is used on a call that is targeted by an `append-args`
advice, the new arguments were simply added to the list of values, which
resulted in invalid code.

To address this, the `append-args` advice now injects an IIFE that
correctly generates a new slice with the additional arguments injected
in. The use of an IIFE allows removing problems if the variadic argument
is provided as a function call (calling that multiple times could result
in duplicated side-effects, etc...).

Fixes #88