fix: panic when passing insufficient arguments to `go` and `toolexec`
Insufficient length checks caused unexpected panics to occur when not
providing any arguments to the `go` subcommand (which ought to be a
valid move), as well as when passing no arguments to the `toolexec`
subcommand (which is not valid use).

Added the necessary checks to produce reasonable error messages instead
of hard-crashing.