feat: required version check is now non-failing
In order to simplify the onboarding experience, it is no longer a hard
requirement to have orchestrion listed in `go.mod`.

If present in `go.mod`, that version will be respected. If not, a
distinctive warning will be printed in the `go version` output, and an
`orchestrion pin` command allows for easy, automated insertion of the
`orchestrion` dependency (adding a `orchestrion.tool.go` file and
running the adequate `go get` command for you).