fix: excessive GOCACHE busting on dev builds (#91)
It turns out binaries installed with `go install` lack VCS information in their build information. This is also true of binaries produced by `go run` and `go test`. This caused the cache busting mechanism to be excessively conservative, resulting in the complete inability to re-use cache objects in the "golden path".
    
To address that, we no longer use most of the VCS information, instead using the `BuildInfo.Main.Version` string, which is set to `(devel)` by `go run`, and the `vcs.modified` `BuildInfo.Settings` entry if `"true"`. If either is true, we try to compute the current binary's SHA512 checusum and embed this in the cache keys; so that re-using the exact same binary should allow re-using GOCACHE entries.