feat: automatic pinning of orchestrion
This adds a new feature which will automatically create the
`orchestrion.tool.go` file, then run `go get` to explicitly install the
current version of `orchestrion` in the `go.mod` file, then run
`go mod tidy` to ensure the `go.mod` file is clean.

This is done only if the required version check fails (indicating the
`go.mod` file did not contain a version for `orchestrion`). A new
environment variable is set so that child processes skip the check if it
was successful for their parent, saving time & risk for errors.

In addition to this, a new `orchestrion warmup` command will pre-build
the instrumented standard library as well as all packages `orchestrion`
may be injecting into compiled code, speeding up future compilations.