feat: automatic pinning of orchestrion (#96)
This adds a new feature which will automatically create the
`orchestrion.tool.go` file, then run `go get` to explicitly install the
current version of `orchestrion` in the `go.mod` file, then run `go mod
tidy` to ensure the `go.mod` file is clean.

This is done only if the required version check fails (indicating the
`go.mod` file did not contain a version for `orchestrion`). A new
environment variable is set so that child processes skip the check if it
was successful for their parent, saving time & risk for errors.

Upon automatic pinning, the following message is print out to STDERR
(formatting varies depending on whether it's a PTY or not, etc... and if
it's a PTY, the box width is ajdusted to fit the viewport):
```
╭──────────────────────────────────────────────────────────────────────────────╮
│                                                                              │
│  Warning: github.com/datadog/orchestrion is not present in your go.mod       │
│  file.                                                                       │
│  In order to ensure build reliability and reproductibility, orchestrion      │
│  will now add itself in your go.mod file by:                                 │
│                                                                              │
│      1. creating a new file named orchestrion.tool.go                        │
│      2. running go get github.com/datadog/orchestrion@v0.7.0-dev.2           │
│      3. running go mod tidy                                                  │
│                                                                              │
│  You should commit the resulting changes into your source control system.    │
│                                                                              │
╰──────────────────────────────────────────────────────────────────────────────╯
```

Fixes #95
Fixes #87 

---

In addition to this, a new `orchestrion warmup` command will pre-build
the instrumented standard library as well as all packages `orchestrion`
may be injecting into compiled code, speeding up future compilations.