fix: line numbering glitch (#109)
In some circumstances, line numbering preservation could miss emitting a
directive when a regular node is nested within a synthetic node (i.e: at
a function call parameter location). This fixes that, and also tweaks
blank line insertion to better preserve the original file layout.

NB - this isn't tested directly here but was designed to address an
issue identified when writing tests for #103.