refactor: replace inject-source-file with inject-declarations (#115)
After discovering it is extremely painful to track down & properly
handle porting imports from the injected source file into the modified
file; and realizing we have full featured support in the `code.Template`
object, replaced the `inject-source-file` advice with a new
`inject-declarations` version, which accepts a template, which is
strictly more flexible.

It also supports specifying a list of `links`, which is packages that
need to be linked against to resolve any `//go:linkname` directive
within the injected declarations; and specifying any import path in
there implies `unsafe` needs to be imported.

---------

Co-authored-by: Eliott Bouhana <47679741+eliottness@users.noreply.github.com>