fix: linker may fail to satisfy linkdeps unless imported by main (#117)
Go's `linker` does not bring in all objects from the `importcfg` file
unconditionally, instead it only loads objects (transitively)
**imported** by the `main` module. This means that _linkDeps_ introduced
to satisfy `//go:linkname` directives may not be satisfied by the linker
(resulting in undefined symbol errors) unless those dependencies are
imported.

This change adds an `OnCompileMain` transformer to the `Weaver` that
adds a new synthetic source file that contains blank imports of any
yet-to-be-satisfied _linkDeps_ entry, ensuring the linker later
satisfies them. It also has the benefit of making sure the compiler
correctly registers any `init` function from these dependencies on the
start-up stack.

This ought to be safe to do in `main`, since this cannot possibly
introduce circular dependencies (and is generally considered good
practice if you're using `//go:linkname` anyway).

---

NB - This was discovered as it broke #108, which will hence contribute
to creating testing coverage for this change.