fix: missing `link.deps` records cause link-time error (#125)
In cases when an instrumented dependency is only present in the final
binary due to injected dependencies, the `link.deps` file might not
contain all necessary information for the linker phase to succeed.

This change adds the necessary logic to properly collect `link.deps`
into all archives so they contain the entire closure of link-time
dependencies.

Also adds an integration test to validate this is correctly working, and
fixes the bug where Orchestrion was not properly loading test files for
decoration.