chore: workflow for automated release creation (#147)
Adds a new workflow that triggers on modifications to the
`internal/version.go` file, and determines whether a new release should
be created or not.

If the version tag does not match an existing release:
- on `pull_request`, requires the PR title be set to `release: <tag>`
- on `push`, automatically creates a _draft_ release with the
corresponding tag

This allows automating release creation by simply merging a pull request
that updates the `Tag` constant in `internal/version.go` and limits the
risk of incorrectly cutting releases.

---------

Co-authored-by: Eliott Bouhana <47679741+eliottness@users.noreply.github.com>