doc: update integration test instructions (#169)
### What does this PR do?

The integration test README.md is stale, referencing scripts and test
structures that no longer exist. Update it to reflect the current way
tests are created and run.

### Motivation

I was trying out Orchestrion and wanted to run the integration tests,
but the instructions were stale. I figured out how to run them based
on the CI scripts.

### Reviewer's Checklist

- [ ] Changed code has unit tests for its functionality.