docs: introduce contributor guide (#183)
Introduces a new contributor guide that provides documentation for the
Aspects syntax, code template language, and provides some guidelines.

This also introduces a JSON schema object that describes the expected
schema of aspects configuration, and configuration files are now
validated by the built-ins generator. The schema also contains the
elements of documentation for join points & advice that is rendered on
the documentation site.

---------

Signed-off-by: github-actions on behalf of RomainMuller <github-actions@github.com>
Co-authored-by: github-actions on behalf of RomainMuller <github-actions@github.com>