chore: dedicated GitHub Workflow to keep go.mod/go.sum up-to-date (#202)
Instead of using Dependabot, which does inspects and udpates each
`go.mod` file separately, issuing one PR for each file; use a custom
GitHub Workflow to update all three `go.mod` files in a single PR.

This also includes provisions to ignore `v1.999.*` releases of the
tracer library, which are experimental releases (that are likely to be
retracted in the future).

This also introduces requirement to pin external actions to the relevant
full-length-SHA, as per OpenSSF recommendations.