ci: setup reusable workflow for integration tests (#158)
<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Commit and PR titles should be prefixed with the general area of the
pull request's change.

-->
### What does this PR do?

This PR creates a new workflow because we can't use the codecov workflow
in dd-trace-go. Even if we are not technically using it because the `if`
condition would have evaluated to false. cf. [this
job](https://github.com/DataDog/dd-trace-go/actions/runs/9892772923)

### Motivation

Add a CI connection dd-trace-go and orchestrion to make sure we don't
break orchestrion features when changing dd-trace-go

---------

Signed-off-by: Eliott Bouhana <eliott.bouhana@datadoghq.com>