internal/injector/aspect/join: only compile Directive regex once
The Directive join point uses a regex to check whether the directive for
a node matches the desired directive. Currently this regex is compiled
every time the Directive is evaluated against a node. This shows up as a
pretty big hot spot in CPU profiles of Orchestrion. We only need to
compile the regex one time, when the Directive is created.

We could do this without a regex if we want to squeeze even more
overhead out of it.