feat: orchestrion job server to reduce duplicated effort (#176)
A new integrated job-server using the NATS protocol is used to share
effort between all involved processes. This reduces the amount of
duplicated work that may be going on; while remaining auditable.

The job-server is a simple request-response server. A server is started
automatically when relevant, listening to a randomly chosen available
port on the loopback interface (via address `127.0.0.1`):

- When using `orchestrion go [build|list|...]`, the job-server is
started by the root `orchestrion` process and shared to all children via
a dedicated environment variable;
- When using `-toolexec="orchestrion toolexec"` (either directly or via
`GOFLAGS`), a new job-server is started by the first `toolexec`
invocation that needs one; and the connect URL is stored in a specific
file within the `go` toolchain's `$WORK` directory, so that other
`toolexec` processes can discover and use it. Such servers automatically
shut down once the `$WORK` directory is destroyed, or if they have not
had any client connected for a set time (`15` minutes).

Job servers maintain a cache of responses provided to requests, and only
compute a given result once. As of this PR, the job server handles:

- Computing the version information added to `-V=full` command
invocations, although this can only benefit from a shared job server if
the build is started by `orchestrion go`, as otherwise these commands
are invoked _before_ the `$WORK` directory is created;
- Resolving injected packages' export files, which is a common source of
duplicated effort, as the `go toolchain` did not know about these and
hence did not account for these dependencies when building the job tree,
such that multiple concurrent resolutions may be attempted.