feat: add profiling of Orchestrion (#217)
### What does this PR do?

Add hidden flags to enable profiling Orchestrion. Options for writing
CPU and heap profiler, as well as execution traces, with a configurable
path prefix. Enabling the profiler for the parent enables it for the
children. There is one of each enabled profile type per process, and
each file gets the PID in its name.

Note that CPU profiling adds a ~200ms delay at the end of the app to
stop profiling. This might matter more for Orchestrion since there are
many normally short-lived processes.

Execution tracing should be enabled with caution since traces can be
quite large. Even the other profile types can produce significant
amounts of data. The profiles will have redundant symbol data. The CPU
and heap profiles can be combined into a single profile to eliminate the
redundancy.

Example usage:

```
$ mkdir profiles
$ orchestrion --profile-prefix=${PWD}/profiles --cpu-profiling=true go build
# To combine the profiles:
$ go tool pprof -proto profiles/*.pprof > cpu.pprof
# To view:
$ go tool pprof -http=localhost:6060 cpu.pprof
```

### Motivation

Looking for hot spots & bottlenecks in Orchestrion

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no
problems
during the review but the signing off is to be done by the reviewer(s).
-->

- [ ] Changed code has unit tests for its functionality.

---------

Co-authored-by: Romain Marcadier <romain@datadoghq.com>