Apply //dd:orchestrion-enabled to tracer-internal modules
The //dd:orchestrion-enabled directive aspect is currently not applied
to tracer-internal modules. dd-trace-go will rely on this directive to
detect whether it was built with Orchestrion. So, set "tracer-internal:
true" for the directive.

I happened to notice this was broken while trying to add
Orchestrion-related information to the profiler uploads. The
`internal/orchestrion.Enabled` function in dd-trace-go was always
returning false, even when built with Orchestrion.