fix: propagate --log-level and --log-file to child processes
If the log level is set via `--log-level`, set the corresponding
environment variables so that the child processes will get it. This was
done for `--log-file` already, but only if it was originally a relative
path. Propagate it always.

I noticed this when investigating #223. I was setting the log level to
"trace" via the command line and didn't see much output. It took me a
bit to realize that the setting wasn't picked up by all the workers. I
set the env var instead, but ideally the command line argument works the
same way :)