feat: support `//dd:span` directive on context-less functions
When building with `orchestrion`, the tracer library is able to access
trace context from the GLS, making it possible to correctly trace
functions that do not receive a context parameter.

This PR changes how the `//dd:span` instrumentation is performed so that
it no longer requires a `context.Context` value to be available
(possibly via a `net/http.Request`), and instead passes a
`context.TODO()` so that `tracer.StartSpanFromContext` produces a span
using GLS data.

---

BREAKING CHANGE: The `.FindArgument` template function has been replaced
by the `.Function.ArgumentOfType` method. A new `.Function.ResultOfType`
method was added to match the same semantics on result values. The
`.Function.Returns` method was renamed to `.Function.Result`.