fix(goflags): parse from `$GOFLAGS` and fixup `-coverpkg`
In order to properly convey coverage flags to child processes when doing
injected package resolution, it is important the `-cover` flags only
applies to packages selected for instrumentation by the user.

This implies we must make the default value of `-coverpkg` explicit if
it is not provided by the user.