refactor(injector): new API with better performance
Create a new `injector.Injector` API that does not rely on
`decorator.Load` (internally using `packages.Load`), instead using the
basic `go/types` API to type-check the AST nodes in order to obtain the
`Uses` map that is sufficient to build an import-managing
`decorator.Decorator` instance.

The package name resolution is done by parsing type information from the
archives mentioned in the `importcfg` file.

This change removes unnecessary compilation of un-instrumented archives
which are not useful, and hence saves time and disk space.