fix(goflags): using `-cover` causes fingerprint mismatch (#230)
In order to properly convey coverage flags to child processes when doing
injected package resolution, it is important the `-cover` flags only
applies to packages selected for instrumentation by the user, otherwise
some packages may be instrumented or not instrumented depending on
whether they are directly resolved or a transitive dependency of a
directly resolved package; and this leads to link-time fingerprint
mismatches.

This implies we must make the default value of `-coverpkg` explicit if
it is not provided by the user.