fix: GLS slot not cleaned up after goroutine exits (#232)
### What does this PR do?

Specifically `nil` out the GLS slot when the goroutine exits, so that
the underlying map can be GC'd and it is not incorrectly re-used by the
next goroutine if that g is recycled from the gf list.

### Motivation

The GLS slot was not reset to nil when the goroutine exits, causing the
old context to be re-
used whenever the g object was recycled from the free list.