fix: apply `//dd:orchestrion-enabled` to tracer-internal modules (#223)
### What does this PR do?

The //dd:orchestrion-enabled directive aspect is currently not applied
to
tracer-internal modules. dd-trace-go will rely on this directive to
detect
whether it was built with Orchestrion. So, set "tracer-internal: true"
for the
directive.

Nothing seems to be broken right now. I'm not sure the best way to test
this, at
least in this PR. I'm working on an integration test for the profiler
when
built with Orchestrion. Once that's up and running, it will break if the
profiler relies on the directive to add Orchestrion-related info. And
presumably once we start relying on goroutine-local storage for context,
stuff
will break if the directive doesn't work in dd-trace-go?

### Motivation

I happened to notice this was broken while trying to add
Orchestrion-related
information to the profiler uploads. The `internal/orchestrion.Enabled`
function in dd-trace-go was always returning false, even when built with
Orchestrion.

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no
problems
during the review but the signing off is to be done by the reviewer(s).
-->

- [ ] Changed code has unit tests for its functionality.

---------

Co-authored-by: Romain Marcadier <romain@datadoghq.com>