refactor(injector): new API with better performance (#234)
Create a new `injector.Injector` API that does not rely on
`decorator.Load` (internally using `packages.Load`), instead using the
basic `go/types` API to type-check the AST nodes in order to obtain the
`Uses` map that is sufficient to build an import-managing
`decorator.Decorator` instance.

The package name resolution is done by parsing type information from the
archives mentioned in the `importcfg` file.

This change removes unnecessary compilation of un-instrumented archives
which are not useful, and hence saves time and disk space.

---

The new API also does not consider `PreserveLineInfo` to be optional
(this was never exposed to end-users anyway), so a bunch of test
reference files have changed to now include line directives.

---------

Co-authored-by: Eliott Bouhana <47679741+eliottness@users.noreply.github.com>