chore: improve CodeCov configuration (#247)
Upload all reports to CodeCov at once at the end of the validation
workflow, so that it stops producing mid-flight coverage reports that
don't account for the entire set of reports.

Also, configure coverage collection during integration test execution.

Noticed the linter configuration was outdated and largely inert, so I
fixed it up and applied A TON of linter fixes. The rule set is obviously
subject to discussion & may also be updated in the future.

---------

Co-authored-by: Eliott Bouhana <eliott.bouhana@datadoghq.com>