chore: make array expansions empty-safe (#280)
Bash treats arrays as defined only if at least one index has been
assigned a value, meaning empty arrays are treated as undefined, which
trips when `set -u` is being used. The safe idiom to use in this case is
`${arr[@]+"${arr[@]}"}`, which only tries expanding the array if it is
deemed set. This is not very pretty, but should do the deed.