chore: separate integration tests in its own packages (#277)
Change the behavior of the code generation from `_integration-tests` to
generate individual test files in each test package.

This speeds up significantly when running only one test suite, which is
particularly useful when developing a new integration.

Also, this change is required for
https://github.com/DataDog/orchestrion/pull/273, where we are testing
both `confluent-kafka-go` v1 and v2. Since both of them depend on CGO
`librdkafka`, they cannot coexist in the same package.