fix: CompileCommand may return extraneous files
In cases where an argument that is not positional ends with `.go`, the
`CompileCommand.GoFiles` function might erroneously return that value
as a source file to be compiled. This is for example the case when
compiling the `github.com/nats-io/nats.go` package, as the `-p` argument
has a value ending in `.go`.

To address this, this PR introduces a more reliable arguments parser
that uses a generated parser (built using the standard `flag` package)
based on the usage documentation found in the output of
`go tool <compile|link>`, so that the flags are more clearly identified.

This replaces the old arguments parser, which is no longer used.