fix: case-based name conflict with `DataDog`
Case-conflicting import paths cause issues when using `go mod vendor` as
the layout on case-sensitive file systems is different from that on
case-insensitive file system. This is a long-known issue with the Go
toolchain, however fixing this is non-trivial and could have undesirable
consequences.

Since the capitalized form of `DataDog` (sic) is most common today,
renaming the `orchestrion` module is the most pragramtic solution today.

Caused-By golang/go#38342
Fixes #287