feat: instrument source files in parallel (#302)
Instrumentation of source files can be performed in parallel, as the AST
transformations do not span across files. Once the files are parsed &
type checked, using a distinct decorator per file enables parallel
processing & more efficient post-processing (fewer items in the
decorator maps, none of which irrelevant to the current AST being
post-processed).

This would improve the performance of builds, and a local profile showed
a 70% improvement on the specific package I was playing with.