chore: refactor test setup (#301)
Summary of changes:

- Do not use `-v` flag for tests in the CI: this makes the output
extremely verbose which is rarely useful, since the default behavior is
to show the verbose output for failed tests only, which is what you are
usually interested on.
- Rename `trace.Span` -> `trace.Trace`: this should make naming a little
less confusing (now `ExpectedTraces` returns `trace.Traces` instead of
`trace.Spans`). Clarified in the comment that `trace.Trace` represents
the root span of a trace.
- Change the type of `Meta` to `map[string]string`
- Added a new field `Metrics` and display in a similar way as `Meta` in
the tree.
- Added some verbose logs displaying all the traces / spans received (I
found this useful while working on this PR).
- `mockAgent.NewSession` now runs `t.Cleanup` for proper test cleanup,
so the `defer checkTraces` call is no longer necessary - I was finding
the output of the traces not found assertion not helpful while working
on the actual test to work.

---------

Co-authored-by: Romain Marcadier <romain.muller@telecomnancy.net>