chore: reduce allocations in injector
Use a `sync.Pool` to reduce allocations performed by the injector, and
change the `functionOption` implementations so that they no longer
receive the context as a pointer (which caused a heap escape); instead
by-copy (the struct is 64 bytes, which is small enough).

This removes a non-trivial amount of allocations, which may have
non-negligible performance impact.