chore: allow validation to run for ALL commits on push (#314)
The concurrency group naming results in cancelling concurrent executions
on the same ref on all events, which means individual commits pushed to
`main` may not be fully validated. In particular, this results in
missing coverage information & benchmarking results.

Changed the naming so that on `push` events, the commit SHA is used
instead of the commit ref, allowing individual commits to run
independently from each other.