fix: `//dd:span` on `func main()` does not work
Since the tracer was started at the start of the `main` function, a
`//dd:span` annotation there may not have been honored, depending on the
order in which aspects were woven into the file.

This PR mvoes the tracer initialization to a `func init()` function, so
that it happens strictly before the `main` function starts executing,
making it possible to `//dd:span` the `main` function reliably.