feat: code templates can declare required go langage level (#357)
In certain cases, it is not practical to write aspects using go syntax
that is known to work with all go versions. While this is in general not
an issue (since `orchestrion` only supports the latest 2 go releases),
it can still cause issues when injecting code into modules with a
`go.mod` file that declares an antiquated `go` version (e.g: `go 1.13`).

In such cases, go 1.22 and newer toolchains will instruct the compiler
to process the source code in a way that ensures compatibility with that
antiquated language level, prohibiting use of newer language features.

To protect ourselves from this issue, the `template` objects can now be
configured to demand a certain minimum language level to be available,
which will cause compile units into which that template generated code
to have the compiler's `-lang` flag set to that level (unless it was
already set at a higher level).

This might have consequences on the behavior of the compiled code (e.g,
the change of `for range` variable bindings introduced in go 1.22); but
we currently do not deal with this particular issue, as we assume it
will seldom have any practical impact in our use-cases.