feat: new `built.WithOrchestrion` constant
Expose a `built.WithOrchestrion` constant so that usages where it is
possible to take a direct dependency on `github.com/DataDog/orchestrion`
is possible are simpler; as a readily available variable is provided.

This reduces the amount of places where the `//dd:orchestrion-enabled`
directive is necessary, which may improve compilation speeds (as
directive join points may be dispandious to evaluate).