fix: net/http server-side instrumentation (#403)
Resolves https://github.com/DataDog/orchestrion/issues/173, resolves
https://github.com/DataDog/orchestrion/issues/400

This PR changes `net/http` server-side instrumentation aspects to trace
the library itself (similar approach as we are doing today with the
client-side instrumentation). The hook where we inject is the
`http.Server.Serve` method, which should be the common method that is
always called when the http server is started (and at this point, the
Handler should already be set).

This fixes problems in the existing instrumentation where we:
- Don't trace custom `http.Handler` implementations
(https://github.com/DataDog/orchestrion/issues/173).
- We double-trace `http.HandlerFunc(func(w http.ResponseWriter, r
*http.Request))` declarations.
- We might even break the build (as described in
https://github.com/DataDog/orchestrion/issues/400).
- Patterns using assignment like `srv.Handler = myHandler` are not
traced.

As a side-effect, this approach removes the ability to use
`orchestrion:ignore` comments to skip tracing on http servers (which
happens for other aspects as well where we instrument the library code).

---------

Co-authored-by: Romain Marcadier <romain.muller@telecomnancy.net>
Co-authored-by: Romain Marcadier <romain@datadoghq.com>