fix(injector: lineinfo): line directive without line number is ignored (#457)
Previously, line directives from the `<generated>` file would appear
without line number as `<generated>` because `<generated>:0` would make
the compiler fail with the error: ` invalid line number: 0`. But we
found out that `<generated>` is actually ignored by the go compiler.
This means that we have to set the line number to 1 and let it appear

Signed-off-by: Eliott Bouhana <eliott.bouhana@datadoghq.com>