fix: failed to resolve argv0 ("/run/rosetta/rosetta")
When running orchestrion in a Docker container using a Rosetta-backed
engine (using the `vz` type of VM), the `argv0` found when crawling
process trees is the `binfmt` handler (`rosetta`) instead of the actual
command being executed.

The `rosetta` handler is only present on the hypervisor (hence it fails
on `no such file or directory` during resolution attempts). When
encountering not found errors on `linux/amd64` platforms, where the
binary name is `rosetta`, try to resolve `argv[1]` as a fallback.