chore: exponential backoff in flaky kafka test (#479)
The segmentio_kafka test has been falking due to the test container not
having the topic and partition ready by the time the test tries to
produce messages. This is possiblye due to the test waiting up to 1
second, while the container has been observed to need almost 2 seconds
to transition the initial partition to available.