fix: OnCompileMain needs to register new `link.deps` (#501)
The `OnCompile` hook cannot necessarily resolve the transitive
dependencies of all link-time dependencies, as those are normally used
to avoid creating dependency cycles.

The `OnCompileMain` hook however MUST resolve the transitive
dependencies of everything, as it needs to introduce `import`s of all
synthetic link-time dependencies so that they are correctly registered
(their `init` functions get scheduled, and they get presented to the
linker). This was previously not necessarily done, and could result in
link-time failures.

---

This involved re-working the way the `link.deps` file is added to the
`_pkg_.a` files so that both the `OnCompile` and `OnCompileMain`
functions contribute to the same closure and it gets written &
registered exactly once. As a result `Command.OnClose` was removed (no
longer used).