fix: build fails with "sending never-build-twice request failed"
The never-build-twice feature assumed that it is not, ever, possible for
a single build to result in multiple different builds of the same
package with different build IDs. #507 suggest this might actually be a
valid outcome (the exact details of why/how are not clear yet).

This PR removes that assumption and keeps separate cache entries for
each individual build ID in order to reduce the risk of breaking the
build.

Fixes #507