fix: load built-in aspects from binary instead of resolving YAML (#517)
The built-in orchestrion aspects must be loaded from the running binary
instead of being loaded from disk. This ensures the correct version of
aspects is being used even if the running version is not the one
declared in `go.mod` (which cannot always be the case anyway).

In addition to this, `orchestrion pin` should upgrade `go.mod` to the
running version of `orchestrion` if it's newer than what's already in
there; and we default-install `dd-trace-go.v1@main` for now (pending the
next release).

This change also changes the `go` directive in the `go.mod` files to be
pinned to the `.0` patch level of the current lowest supported go minor,
instead of pinning it to the latest available one (which is unnessary
and can be inconvenient to users).