fix: required-version respawning of orchestrion causes build failures (#522)
The automatic version check incurring a re-spawn of orchestrion with the
`go.mod` version cannot happen in all situations, and this results in
inconsistent build IDs being generated when manually specifying
`-toolexec` (instead of using `orchestrion go ...`).

Since we cannot always do the version check, we cannot re-spawn in all
situations. For now, we decided to no longer re-spawn but instead
produce specific error messages when the running version is not the one
required by `go.mod`, instructing users what to do to fix that.

Fixes #507