fix: job server fails to start on Windows due to locking error (#545)
Updated the `*filelock.Mutex` type to implement `io.ReadWriteSeeker`,
`io.ReaderAt`, and `io.WriterAt` so that it can be used to read/write
the locked file without having to open a secondary file descriptor. This
is necessary on Windows, where the locks are tied to the specific file
descriptor (as opposed to UNIX where the lock is tied to the process).

---------

Co-authored-by: Eliott Bouhana <47679741+eliottness@users.noreply.github.com>