chore: fix broken dependency updater workflow (#549)
The dependency updater workflow "accidentally" worked previously, but no
longer does. The culprit is we were using `jq` without `-r` which was
causing it to output a JSON string (quotes included), instead of just
the version. Previous versions of the toolchain silently accepted this
and un-quoted the string, but newer releases no longer do this and
instead fail.

This problem went un-noticed because the `find` command's exit status
does not account for the exit statuses of commands executed by `-exec`
and `-execdir` (it only exits non-`0` if it encounters an issue while
traversing the directory tree).

This commit uses a `for` loop instead of `-execdir` so that the workflow
"correctly" fails if one of the commands there happens to fail.