offload packages.Load calls to the job server
The config loader spends A LOT of time resolving the same packages over
and over in each process, and offloading this task to the job server
allows caching and re-using results, significantly reducing the amount
of child processes necessary to load the configuration on a hot cache.

This significantly improves the performance of real world builds.