feat: instrument Orchestrion with Datadog tracing (#577)
In order to facilitate investigation of build performance, instrument
Orchestrion (manually, because who does not love some irony) with the
`dd-trace-go` V2 library, so that we can obtain traces from Orchestrion
executions.

This is enabled by setting `ORCHESTRION_TRACE` to any non-blank value,
and requires a Datadog Agent to be available and appropriately
configured in order for traces to be sent to the backend.

Child process distributed tracing is achieved by using environment
variables as a propagation medium, using a custom-made carrier that
works off of `os.Environ()`-style slices.

The NATS client/server is also instrumented for distributed tracing,
using `nats.Msg.Header` as a carrier (this is akin to using HTTP or gRPC
headers for the same purpose).

------------------------------------------------------------------------

Initial investigations of traces (building orchestrion with orchestrion)
showed that a far-from-trivial overhead was induced by an abundance of
sub-process creation to `go tool pack` (involved in storing & accessing
the `link.deps` files). Replacing these with a simple library
(`github.com/blakesmith/ar`) allowed reducing the overhead of the
`Weaver.OnLink` call by a significant amount (shaving 6 seconds off of
this particular build).

---

Jira: APPSEC-57051

---------

Co-authored-by: Eliott Bouhana <47679741+eliottness@users.noreply.github.com>