fix: fingerprint mismatch when using go.work files
When a `go.mod` file contains a `replace` directive for a module that is
part of the current workspace (and directing to the in-workspace
directory), the `go list` command does not return a `Replace` field for
that module; instead it returns a blank `Version` field.

This commit fixes the file fingerprinting decision so that it hashes
the modules' content when the `Version` field is blank or when the
module is replaced by a local directory.

Also added a unit test to validate things work as expected in this case.